Moved validPasswordFormat function in separete file
